Inline headers from demo-rasterizer
